+++ CHANGELOG.md
@@ -5,10 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Python PEP 440 Versioning](https://www.python.org/dev/peps/pep-0440/).
 
 ## [Unreleased]
+### Added
+- Compatibility with RDFLib v7.0.0 - Closes #197
 ### Fixed
 - `sh:qualifiedMinValue` on `sh:qualifiedValueShape` now works again, even if there are no value nodes found
   on the path of the parent `PropertyShape`. Fixes #213 Thank you @ajnelson-nist for finding and reporting this.
+- Fixes in rdfutl (clone dataset, mixin dataset, and innoculate dataset) to support the case where all the DS's
+  triples are in the default-context-uri graph.
 ### Changed
+- In accordance with corresponding changes in RDFLib v7.0.0, PySHACL will now always use the default-context-uri graph
+  when parsing a grpah into a Dataset or a ConjunctiveGraph
 - Switched from deprecated `pkg_resources` to `importlib.metadata` for compatibility with Python 3.11 and 3.12.
   - This changes the way `pyshacl[extras]` are detected at runtime. If this adversely affects you, let us know.
 - Bumped PrettyTable dependency to a much newer version, to fix distro packaging conflicts and other issues.